refactor: test cases for TrendingActorsViewModel
